# public/openapi.yaml
openapi: 3.0.3
info:
  title: HR API Documentation
  version: 1.0.0
servers:
  - url: http://127.0.0.1:8000
    description: Local Development Server
  - url: https://api.yourdomain.com
    description: Production Server (Staging/Live)
...
paths:
...
  /api/departments/{id}:
    patch:
      summary: Update department details
      operationId: updateDepartment
      tags:
        - Departments
      security:
        - bearerAuth: []
      parameters:
        - in: header
          name: Accept
          required: true
          schema:
            type: string
            default: application/json
          example: application/json
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: Department ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                  maxLength: 255
                  example: Engineering & Operations
                description:
                  type: string
                  maxLength: 1000
                  nullable: true
                  example: Handles soft dev and infrastructure.
                manager_id:
                  type: integer
                  nullable: true
                  example: 3
      responses:
        "200":
          description: Department updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Department updated successfully
                  data:
                    type: object
                    properties:
                      id:
                        type: integer
                        example: 1
                      name:
                        type: string
                        example: Engineering & Operations
                      description:
                        type: string
                        nullable: true
                        example: Handles soft dev and infrastructure.
                      status:
                        type: string
                        example: active
                      manager:
                        type: object
                        nullable: true
                        properties:
                          id:
                            type: integer
                            example: 3
                          name:
                            type: string
                            example: Alex Johnson
                          email:
                            type: string
                            example: alex@example.com
                          job_title:
                            type: string
                            example: VP of Engineering
                      employees_count:
                        type: integer
                        example: 12
                      created_at:
                        type: string
                        example: 2026-01-15 10:00:00
              example:
                success: true
                message: Department updated successfully
                data:
                  id: 1
                  name: Engineering & Operations
                  description: Handles soft dev and infrastructure.
                  status: active
                  manager:
                    id: 3
                    name: Alex Johnson
                    email: alex@example.com
                    job_title: VP of Engineering
                  employees_count: 12
                  created_at: 2026-01-15 10:00:00
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: Unauthenticated
                  errors:
                    type: string
                    nullable: true
                    example: null
              example:
                success: false
                message: Unauthenticated
                errors: null
        "403":
          description: Forbidden - User lacks department.edit permission
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: User does not have the right permissions.
                  errors:
                    type: string
                    nullable: true
                    example: null
              example:
                success: false
                message: User does not have the right permissions.
                errors: null
        "404":
          description: Department not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: Department not found
                  errors:
                    type: string
                    nullable: true
                    example: null
              example:
                success: false
                message: Department not found
                errors: null
        "422":
          description: Validation error
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: Validation error
                  errors:
                    type: object
                    additionalProperties:
                      type: array
                      items:
                        type: string
              example:
                success: false
                message: Validation error
                errors:
                  name:
                    - The name has already been taken.
  /api/departments/{id}/change-status:
    patch:
      summary: Toggle department active/inactive status
      operationId: changeDepartmentStatus
      tags:
        - Departments
      security:
        - bearerAuth: []
      parameters:
        - in: header
          name: Accept
          required: true
          schema:
            type: string
            default: application/json
          example: application/json
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: Department ID
      responses:
        "200":
          description: Department status changed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Activated department successfully with keeping employee records.
                  data:
                    type: object
                    properties:
                      id:
                        type: integer
                        example: 1
                      name:
                        type: string
                        example: Engineering
                      description:
                        type: string
                        nullable: true
                        example: Software Development Department
                      status:
                        type: string
                        example: active
                      manager:
                        type: object
                        nullable: true
                      employees_count:
                        type: integer
                        example: 12
                      created_at:
                        type: string
                        example: 2026-01-15 10:00:00
              examples:
                activated:
                  summary: Department Activated
                  value:
                    success: true
                    message: Activated department successfully with keeping employee records.
                    data:
                      id: 1
                      name: Engineering
                      description: Software Development Department
                      status: active
                      manager: null
                      employees_count: 12
                      created_at: 2026-01-15 10:00:00
                deactivated:
                  summary: Department Deactivated
                  value:
                    success: true
                    message: Deactivated department successfully with keeping employee records.
                    data:
                      id: 1
                      name: Engineering
                      description: Software Development Department
                      status: inactive
                      manager: null
                      employees_count: 12
                      created_at: 2026-01-15 10:00:00
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: Unauthenticated
                  errors:
                    type: string
                    nullable: true
                    example: null
              example:
                success: false
                message: Unauthenticated
                errors: null
        "403":
          description: Forbidden - User lacks department.change-status permission
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: User does not have the right permissions.
                  errors:
                    type: string
                    nullable: true
                    example: null
              example:
                success: false
                message: User does not have the right permissions.
                errors: null
        "404":
          description: Department not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: Department not found
                  errors:
                    type: string
                    nullable: true
                    example: null
              example:
                success: false
                message: Department not found
                errors: null
...
  /api/departments/managers-dropdown:
    get:
      summary: Get list of managers for dropdown selection (Allowed for Owner and HR)
      description: Returns a simplified array of all users who can manage departments (ID and Name only) to populate dropdown filters and forms.
      operationId: getManagersDropdown
      tags:
        - Departments
      security:
        - bearerAuth: []
      parameters:
        - in: header
          name: Accept
          required: true
          schema:
            type: string
            default: application/json
          example: application/json
        - in: query
          name: lang
          required: false
          description: Language for response translation (ar, en)
          schema:
            type: string
            enum: [ar, en]
            default: en
          example: en
      responses:
        '200':
          description: Managers list retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Managers list retrieved successfully."
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                          example: 2
                        name:
                          type: string
                          example: "John Manager"
              example:
                success: true
                message: "Managers list retrieved successfully."
                data:
                  - id: 2
                    name: "John Manager"
                  - id: 5
                    name: "Sarah HR"
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '403':
          description: Forbidden (User lacks required permission)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                message: "You do not have the required permissions to perform this action."
                errors: null
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
...
    401Unauthorized:
      description: Unauthenticated
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            success: false
            message: Unauthenticated. Please log in first.
            errors: null
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: An error occurred processing the request.
        errors:
          type: object
          nullable: true
          example: null
...